log. 테스트코드 제거, SatageId 타입변경

+++ src/main/java/org/etmetmy/bn_server/domain/project/dto/request/ProjectStageUpdateRequest.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProjectStageUpdateRequest {
-    private Integer stageId;
+    private Long stageId;
 }
